Hid some cartridges while not reloading
I initially wanted to figure out a way to un-hide the bullet bones when the reload is partially done,
but as far as I can tell, that'd require some baseline edits to ARC-9 which I'm entirely unaware of how I'd even add it.

Currently, the magazine sets itself to visible when a reload starts, but by then the magazine isn't in the players hands yet.
It's also visible when the reload is done, so the ability to hide it after a set amount of time would be great.

lua/weapons/arc9_base/sh_bodygroups.lua, line 149-189